Change CFGAS_BreakPiece::GetLength() to return size_t.
The length should never be negative. Return size_t so there are fewer
locations that mix int and size_t. Add a GetChar() variant that takes
size_t to go with the GetLength() value.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/95130
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ xfa/fde/cfde_textout.cpp
@@ -372,10 +372,10 @@ bool CFDE_TextOut::RetrievePieces(CFGAS_Char::BreakType dwBreakStatus,
   int32_t iCount = m_pTxtBreak->CountBreakPieces();
   for (int32_t i = 0; i < iCount; i++) {
     const CFGAS_BreakPiece* pPiece = m_pTxtBreak->GetBreakPieceUnstable(i);
-    int32_t iPieceChars = pPiece->GetLength();
+    size_t iPieceChars = pPiece->GetLength();
     size_t iChar = *pStartChar;
     int32_t iWidth = 0;
-    int32_t j = 0;
+    size_t j = 0;
     for (; j < iPieceChars; j++) {
       const CFGAS_Char* pTC = pPiece->GetChar(j);
       int32_t iCurCharWidth = std::max(pTC->m_iCharWidth, 0);
